+++ src/app/api/stock/route.ts
@@ -5,9 +5,7 @@ import data from "./data.json";
 export async function GET(request: any) {
   const { searchParams } = new URL(request.url);
   const query = searchParams.get("search");
-  console.log(query);
   if (query === "") return NextResponse.json([], { status: 200 });
-
   const result = data.filter((item) =>
     item.name.toLowerCase().startsWith(query?.toLowerCase() ?? "")
   );
